move /app to /src
