+++ devstart.py
@@ -3,4 +3,4 @@ from dotenv import load_dotenv
 
 if __name__ == "__main__":
     load_dotenv()
-    uvicorn.run("app.main:app", host="localhost", port=8000, reload=True)
+    uvicorn.run("src.main:app", host="localhost", port=8000, reload=True)
