+++ src/difftool/diffsorter.py
