+++ src/difftool/itemdiffer.py
