+++ src/difftool/matcher.py
