+++ src/openapi/strings.py
