+++ src/parsing/cr/refresh_cr.py
@@ -4,14 +4,14 @@ from dataclasses import asdict
 
 import requests
 
-from app.database import operations as ops
-from app.database.db import SessionLocal
-from app.difftool.diffmaker import CRDiffMaker
-from app.parsing.cr import extract_cr
-from app.resources import static_paths as paths
-from app.resources.cache import GlossaryCache, KeywordCache
-from app.utils import notifier
-from app.utils.logger import logger
+from src.database import operations as ops
+from src.database.db import SessionLocal
+from src.difftool.diffmaker import CRDiffMaker
+from src.parsing.cr import extract_cr
+from src.resources import static_paths as paths
+from src.resources.cache import GlossaryCache, KeywordCache
+from src.utils import notifier
+from src.utils.logger import logger
 
 
 def get_response_text(response: requests.Response) -> str | None:
