+++ src/parsing/utils/download_doc.py
@@ -3,7 +3,7 @@ from typing import Literal
 
 import requests
 
-from app.resources import static_paths as paths
+from src.resources import static_paths as paths
 
 
 def download_doc(link: str, kind: Literal["mtr", "ipg"]):
