+++ src/routers/rule_deprecated.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, Path, Query, Response
 from fastapi.responses import FileResponse
 from sqlalchemy.orm import Session
 
-from app.openapi.no422 import no422
+from src.openapi.no422 import no422
 
 from ..database import operations as ops
 from ..database.db import get_db
